fix: [VPNLINUX-1369] don't update tracked processes when changing ST config
We decided it's not worth it to update tracked processes when removing exe paths from ST config.
The reason is that it adds complexity and it's not guarantee that processes that matched
the removed exe paths won't be split tunneled anymore, because they might keep old sockets
alive that are still being split tunneled.